Style: [Client] #9 DownNav style separation
-DownNav에 있는 스타일 styled Nav로 옮김
