+++ Client/src/styled/Nav.tsx
@@ -3,18 +3,18 @@ import { signInClickProp } from '../state/typeDefs';
 
 export const Container = styled.nav`
   position: fixed;
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
   max-width: 768px;
   height: 5rem;
   background-color: #007580;
   width: 100%;
-  padding: 30px;
 `;
 
 export const UpperContainer = styled(Container)`
   top: 0;
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+  padding: 30px;
 `;
 
 export const Selection = styled.select`
@@ -69,4 +69,26 @@ export const LoginButton = styled.button<signInClickProp>`
   border: 0;
   background-color: transparent;
   cursor: grab;
+`;
+
+interface isActiveProp {
+  isActive: boolean;
+}
+
+//DownNav
+export const DownContainer = styled(Container)`
+  bottom: 0;
+`;
+export const NavMenu = styled.ul`
+  display: flex;
+  align-items: center;
+  justify-content: space-around;
+  padding: 1.5rem 0;
+  margin: 0;
+`;
+export const NavMenuList = styled.li<isActiveProp>`
+  font-size: 2rem;
+  color: ${(props) => (props.isActive ? '#fed049' : '#d8ebe4')};
+  display: flex;
+  cursor: pointer;
 `;
